cat_picture_api/requirements.txt

Django==3.2.24
django-rest-auth==0.9.5
djangorestframework==3.14.0
Pillow==8.4.0
requests==2.27.1

